requirements.txt
Flask==2.3.2
pymongo==4.6.1
flask-login==0.6.2
flask-mail==0.9.1
python-dotenv==1.0.0
werkzeug==2.3.7
Pillow==10.0.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
celery==5.3.4
redis==5.0.1
sendgrid==6.10.0
twilio==8.10.0
pytz==2023.3
ics==0.7.2
markdown==3.5.1
bleach==6.1.0